# docs/swagger.yaml
definitions:
...
  domain.Ticket:
    properties:
      conversation_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      description:
        type: string
      id:
        type: string
      priority:
        type: integer
      status:
        type: string
      subject:
        type: string
      tenant_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  domain.User:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      roles:
        items:
          type: string
        type: array
    type: object
...
  models.UpdateTicketStatusRequest:
    properties:
      status:
        enum:
        - open
        - in_progress
        - closed
        type: string
    required:
    - status
    type: object
  models.UserCreateRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      repeat_password:
        type: string
      roles:
        items:
          type: string
        type: array
    required:
    - email
    - name
    - password
    - repeat_password
    - roles
    type: object
  models.UserUpdateRequest:
    properties:
      email:
        type: string
      name:
        type: string
      roles:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
paths:
...
  /api/v1/backoffice/users:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of users with optional filtering by name
        and email.
      parameters:
      - description: Search term for name or email
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      summary: Get a list of users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided details.
      parameters:
      - description: User creation request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      summary: Create a new user
      tags:
      - Users
  /api/v1/backoffice/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by their unique ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieve a user by their unique ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update an existing user's details by their unique ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      summary: Update an existing user
      tags:
      - Users
...
  /tickets:
    get:
      consumes:
      - application/json
      description: Get paginated list of tickets with optional filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by status (open, closed, assigned)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Ticket'
      summary: Get list of tickets
      tags:
      - Tickets
  /tickets/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of a ticket by its ID
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      - description: New status for the ticket
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTicketStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Ticket'
      summary: Update ticket status
      tags:
      - Tickets
swagger: "2.0"